app/requirements.txt
pyserial
websockets
mysql-connector-python
